+++ lib/garner.rb
@@ -1,4 +1,8 @@
+require 'multi_json'
+# garner
 require 'garner/version'
+# objects
+require 'garner/objects/etag.rb'
 # middleware
 require 'garner/middleware/base.rb'
 require 'garner/middleware/cache/bust.rb'