+++ CHANGELOG.md
@@ -2,5 +2,7 @@
 ---
 
 * Initial implementation based on [@macreery](https://github.com/macreery)'s original code.
+* Rack middleware for cache busting, `Garner::Middleware::Cache::Bust` - [@dblock](https://github.com/dblock).
+* Generating ETags, `Garner::Objects::Etag` - [@dblock](https://github.com/dblock).